// frontend/src/app/core/api/model/productDto.ts
/**
 * Gila E-Commerce API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDto { 
    id?: string;
    name: string;
    description?: string;
    price: number;
    stock: number;
    category: string;
}