// frontend/src/app/core/api/model/productImportStatusDto.ts
/**
 * Gila E-Commerce API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductImportStatusDto { 
    taskId: string;
    status: string;
    totalRows: number;
    processedRows: number;
    errorCount: number;
    warnings: Array<string>;
}